+++ cuda/SciFi/classifiers/src/TMVA_Forward.cu
@@ -8,8 +8,8 @@ __host__ __device__ void SciFi::Tracking::Transform_1(float* iv, const TMVA* tmv
   for (int ivar = 0; ivar < nVar; ivar++) {
     const float offset = tmva->fMin_1[cls][ivar];
     const float scale =
-      1.0 / (tmva->fMax_1[cls][ivar] - tmva->fMin_1[cls][ivar]); // TODO speed this up. but then not easy to update :(
-    iv[ivar] = (iv[ivar] - offset) * scale * 2. - 1.;
+      1.0f / (tmva->fMax_1[cls][ivar] - tmva->fMin_1[cls][ivar]); // TODO speed this up. but then not easy to update :(
+    iv[ivar] = (iv[ivar] - offset) * scale * 2.f - 1.f;
   }
 }
 
@@ -22,7 +22,7 @@ __host__ __device__ float SciFi::Tracking::ActivationFnc(float x)
 __host__ __device__ float SciFi::Tracking::OutputActivationFnc(float x)
 {
   // sigmoid
-  return 1.0 / (1.0 + exp(-x));
+  return 1.0f / (1.0f + expf(-x));
 }
 
 __host__ __device__ float SciFi::Tracking::GetMvaValue__(const float* inputValues, const TMVA* tmva)
